Move spec_set_transforms.py to the specs module.
